added support to postproc to create outputs from an entire wrfout file, also runs in standalone mode

+++ postproc.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import matplotlib as mpl
 mpl.use('AGG')
 import matplotlib.pyplot as plt
@@ -95,23 +97,12 @@ class Postprocessor(object):
         uw, vw = get_wisdom(u_name), get_wisdom(v_name)
         u = uw['retrieve_as'](d, tndx)
         v = vw['retrieve_as'](d, tndx)
-        u_lat, u_lon = uw['grid'](d)
-        v_lat, v_lon = vw['grid'](d)
 
-        if u.shape != u_lat.shape:
+        if u.shape != lat.shape:
             raise PostprocError("Variable %s size does not correspond to grid size: var %s grid %s." % (u_name, u.shape, u_lat.shape))
         
-        if v.shape != v_lat.shape:
+        if v.shape != lat.shape:
             raise PostprocError("Variable %s size does not correspond to grid size." % v_name)
-
-        # this is a HACK - instead of proper interpolation, we just average neighbors in staggered dimension
-        u_i = np.zeros_like(lat)
-        for i in range(u_i.shape[1]):
-            u_i[:,i] = 0.5 * (u[:,i] + u[:,i+1])
-
-        v_i = np.zeros_like(lat)
-        for i in range(v_i.shape[0]):
-            v_i[i,:] = 0.5 * (v[i,:] + v[i+1,:])
 
         # look at mins and maxes
         fa_min,fa_max = min(np.nanmin(u), np.nanmin(v)),max(np.nanmax(u), np.nanmax(v))
@@ -125,9 +116,9 @@ class Postprocessor(object):
             v[v < fa_min] = fa_min
             v[v > fa_max] = fa_max
 
-        # create the raster & get coordinate bounds, HACK to get better barbs resolution
-        s = 2
-        raster_png_data,corner_coords = basemap_barbs_mercator(u_i[::s,::s],v_i[::s,::s],lat[::s,::s],lon[::s,::s])
+        # create the raster & get coordinate bounds, HACK to get better quiver resolution
+        s = 3
+        raster_png_data,corner_coords = basemap_barbs_mercator(u[::s,::s],v[::s,::s],lat[::s,::s],lon[::s,::s])
 
         return raster_png_data, corner_coords
 
@@ -157,7 +148,28 @@ class Postprocessor(object):
 
         return raster_path, colorbar_path, corner_coords
 
-    def _scalar2kmz(self, d, var, tndx, out_path):
+
+    def _vector2png(self, d, var, tndx, out_path):
+        """
+        Postprocess a single vector variable ``var`` and stores result in a raster file.
+
+        :param d: the open netCDF file
+        :param var: the variable name
+        :param tndx: the temporal index of the grid to store
+        :param out_path: the path to the KMZ output
+        :return: the path to the raster and the bounding coordinates
+        """
+        # render the raster & colorbar
+        raster_png_data, corner_coords = self._vector2raster(d, var, tndx)
+
+        raster_path = out_path + '-raster.png'
+        with open(raster_path, 'w') as f:
+            f.write(raster_png_data)
+
+        return raster_path, corner_coords
+
+
+    def _scalar2kmz(self, d, var, tndx, out_path, cleanup = True):
         """
         Postprocess a single raster variable ``fa`` and store result in out_path.
 
@@ -165,13 +177,14 @@ class Postprocessor(object):
         :param var: the variable name
         :param tndx: the temporal index of the grid to store
         :param out_path: the path to the KMZ output
+        :param cleanup: if true, delete png files
         :return: the path to the generated KMZ
         """
         # construct kml file
         doc = kml.Kml(name = var)
 
-        # render the raster & colorbar
-        raster_png_data, corner_coords, cb_png_data = self._scalar2raster(d, var, tndx)
+        # generate the png files
+        raster_path, cb_path, corner_coords = self._scalar2png(d, var, tndx, out_path)
 
         # add colorbar to KMZ
         cbo = doc.newscreenoverlay(name='colorbar')
@@ -180,33 +193,28 @@ class Postprocessor(object):
         cbo.size = kml.Size(x=150,y=300,xunits=kml.Units.pixel,yunits=kml.Units.pixel)
         cbo.color = kml.Color.rgb(255,255,255,a=150)
         cbo.visibility = 1
-        cb_name = out_path + "-cb.png"
-        with open(cb_name, 'w') as f:
-            f.write(cb_png_data)
-        doc.addfile(cb_name)
-        cbo.icon.href=cb_name
+        doc.addfile(cb_path)
+        cbo.icon.href=cb_path
 
         # add ground overlay
         ground = doc.newgroundoverlay(name=var,color='80ffffff')
-        raster_name = out_path + '-raster.png'
         ground.gxlatlonquad.coords = corner_coords
-        with open(raster_name,'w') as f:
-            f.write(raster_png_data)
-        doc.addfile(raster_name)
-        ground.icon.href = raster_name
+        doc.addfile(raster_path)
+        ground.icon.href = raster_path
 
         # build output file
         kmz_path = out_path + ".kmz"
         doc.savekmz(kmz_path)
 
         # cleanup
-        os.remove(cb_name)
-        os.remove(raster_name)
+        if cleanup:
+            os.remove(cb_name)
+            os.remove(raster_name)
         
-        return kmz_path
+        return kmz_path, raster_path, cb_path, corner_coords
 
 
-    def _vector2kmz(self, d, var, tndx, out_path):
+    def _vector2kmz(self, d, var, tndx, out_path, cleanup = True):
         """
         Postprocess a single vector variable ``var`` and store result in out_path.
 
@@ -219,26 +227,24 @@ class Postprocessor(object):
         # construct kml file
         doc = kml.Kml(name = var)
 
-        # render the raster & colorbar
-        raster_png_data, corner_coords = self._vector2raster(d, var, tndx)
+        # generate the png files
+        raster_path, corner_coords = self._vector2png(d, var, tndx, out_path)
 
         # add ground overlay
         ground = doc.newgroundoverlay(name=var,color='80ffffff')
-        raster_name = out_path + '-raster.png'
         ground.gxlatlonquad.coords = corner_coords
-        with open(raster_name,'w') as f:
-            f.write(raster_png_data)
-        doc.addfile(raster_name)
-        ground.icon.href = raster_name
+        doc.addfile(raster_path)
+        ground.icon.href = raster_path
 
         # build output file
         kmz_path = out_path + ".kmz"
         doc.savekmz(kmz_path)
 
         # cleanup
-        os.remove(raster_name)
+        if cleanup:
+            os.remove(raster_name)
         
-        return kmz_path
+        return kmz_path, raster_path, corner_coords
 
 
     def vars2kmz(self, wrfout_path, dom_id, ts_esmf, vars):
@@ -264,10 +270,10 @@ class Postprocessor(object):
             try:
                 outpath_base = os.path.join(self.output_path, self.product_name + ("-%02d-" % dom_id) + ts_esmf + "-" + var) 
                 kmz_path = None
-                if var in ['WINDSPD']:
-                    kmz_path = self._vector2kmz(d, var, tndx, outpath_base)
+                if var in ['WINDVEC']:
+                    kmz_path,_,_,_= self._vector2kmz(d, var, tndx, outpath_base)
                 else:
-                    kmz_path = self._scalar2kmz(d, var, tndx, outpath_base)
+                    kmz_path,_,_,_ = self._scalar2kmz(d, var, tndx, outpath_base)
                 kmz_name = osp.basename(kmz_path)
                 self._update_manifest(ts_esmf, var, { 'kml' : kmz_name })
             except Exception as e:
@@ -296,8 +302,8 @@ class Postprocessor(object):
         for var in vars:
             try:
                 outpath_base = os.path.join(self.output_path, self.product_name + ("-%02d-" % dom_id) + ts_esmf + "-" + var) 
-                if var in ['WINDSPD']:
-                    raster_path, coords = self._vector2png(d, var, tndx, output_base)
+                if var in ['WINDVEC']:
+                    raster_path, coords = self._vector2png(d, var, tndx, outpath_base)
                     raster_name = osp.basename(raster_path)
                     self._update_manifest(ts_esmf, var, { 'raster' : raster_name, 'coords' : coords})
                 else:
@@ -306,6 +312,41 @@ class Postprocessor(object):
                     self._update_manifest(ts_esmf, var, { 'raster' : raster_name, 'colorbar' : cb_name, 'coords' : coords})
             except Exception as e:
                 logging.warning("Exception %s while postprocessing %s for time %s" % (e.message, var, ts_esmf))
+
+
+    def process_file(self, wrfout_path, var_list):
+        """
+        Process an entire file, all timestamps.
+
+        :param wrfout_path: the wrfout to process
+        :param var_list: list of variables to process
+        """
+        # open the netCDF dataset
+        d = nc4.Dataset(wrfout_path)
+
+        # extract ESMF string times and identify timestamp of interest
+        times = [''.join(x) for x in d.variables['Times'][:]]
+
+        # build one KMZ per variable
+        for tndx, ts_esmf in enumerate(times):
+            print('Processing time %s ...' % ts_esmf)
+            for var in var_list:
+                try:
+                    outpath_base = os.path.join(self.output_path, self.product_name + '-' + ts_esmf + '-' + var) 
+                    if var in ['WINDVEC']:
+                        kmz_path,raster_path,coords = self._vector2kmz(d, var, tndx, outpath_base, cleanup=False)
+                        raster_name = osp.basename(raster_path)
+                        kmz_name = osp.basename(kmz_path)
+                        self._update_manifest(ts_esmf, var, { 'raster' : raster_name, 'coords' : coords, 'kml' : kmz_name})
+                    else:
+                        kmz_path,raster_path,cb_path,coords = self._scalar2kmz(d, var, tndx, outpath_base, cleanup=False)
+                        raster_name, cb_name = osp.basename(raster_path), osp.basename(cb_path)
+                        kmz_name = osp.basename(kmz_path)
+                        self._update_manifest(ts_esmf, var, { 'raster' : raster_name, 'colorbar' : cb_name, 'coords' : coords, 'kml' : kmz_name })
+                except Exception as e:
+                    logging.warning("Exception %s while postprocessing %s for time %s" % (e.message, var, ts_esmf))
+
+
 
 
     def _update_manifest(self,ts_esmf, var, kv):
@@ -324,4 +365,23 @@ class Postprocessor(object):
         vd.update(kv)
         mf_path = os.path.join(self.output_path, self.product_name + '.json')
         json.dump(self.manifest, open(mf_path, 'w'))
+
+
+
+
+if __name__ == '__main__':
+    
+    if len(sys.argv) != 4:
+        print('usage: %s <wrfout_path> <var_list> <prefix>' % sys.argv[0])
+        sys.exit(1)
+
+    wrf_path = sys.argv[1]
+    var_list = sys.argv[2].split(',')
+    prefix = sys.argv[3]
+
+    p = Postprocessor(os.path.dirname(prefix), os.path.basename(prefix))
+    p.process_file(wrf_path, var_list)
+
+
+
